refactor utils and types of token creator
